Dealt with SerializationTable lifetime issue by generating the table inside the serialization function
This makes the serialization interface the same as the deserialization
interface for users. Deserialization already had to regenerate the table
each time, because arbitrary crates cannot necessarily know the concrete
type of the MapVisitor other than inside the visit_map() function.

It feels like we shouldn't have to regenerate the table every time, but
the runtime cost is trivial – there will generally be FAR more fields
than columns.