Replace the error limiting code
The old system limited the amount that the actual-transfer-rate could
diverge from the ideal-transfer-rate. In order to avoid cumulative error
over long periods of time, it tried to limit the duration between
updates to a quarter second, but the system it used to do that was
broken.

The new system (approximately) puts a fixed limit on the CUMULATIVE
error that accumulates between any 2 consecutive updates. This still
doesn't converge to a sensible-looking state, but it's less weird.