Use VecDeque for SimpleTimeline so forget_before() can properly drop all possible old event handles
Since I had to adapt binary search to VecDeque anyway, I took this time
to adopt "binary search from the back" for search_changes(), so that the
cost is O(log(number of changes after the when you're searching for))
instead of O(log(number of changes in the entire history)).

These changes appeared to cause a minor pessimization of
search_changes() in the current handshakes tests, and a larger minor
optimization of forget_before(). The main point was to reduce memory
leakage, which I haven't been testing.